requirements.txt
matplotlib
numpy
pandas
pandas_ta
pandas-datareader
# datatable
joblib
Flask
pqdm
inquirer
honcho
boto3
fsspec
gcsfs
google-cloud-storage
google-cloud-logging
google-api-python-client
oauth2client
requests
ccxt
websocket-client
python-binance
gunicorn
schedule
timeout_decorator
rich
discord
python-dotenv